+++ PidginMacOSItegration/Gtk/GtkWindow.h
@@ -1,0 +1,16 @@
+//
+//  GtkWindow.h
+//  PidginMacOSIntegration
+//
+//  Created by Agalin on 27/07/2020.
+//  Copyright © 2020 Agalin. All rights reserved.
+//
+
+#ifndef GtkWindow_h
+#define GtkWindow_h
+
+#include <gtk/gtk.h>
+
+GtkWindow* to_window(void* obj);
+
+#endif /* GtkWindow_h */
