+++ PidginMacOSItegration/Menu.h
@@ -1,0 +1,16 @@
+//
+//  Menu.h
+//  PidginMacOSIntegration
+//
+//  Created by Agalin on 27/07/2020.
+//  Copyright © 2020 Agalin. All rights reserved.
+//
+
+#ifndef Menu_h
+#define Menu_h
+
+#include <gtk/gtk.h>
+
+void set_menu(GtkWidget *menu);
+
+#endif /* Menu_h */
