+++ PidginMacOSItegration/Pidgin/PidginImage.h
@@ -1,0 +1,19 @@
+//
+//  PidginImage.h
+//  PidginMacOSIntegration
+//
+//  Created by Agalin on 27/07/2020.
+//  Copyright © 2020 Agalin. All rights reserved.
+//
+
+#ifndef PidginImage_h
+#define PidginImage_h
+
+#include <gtk/gtk.h>
+
+struct im_image_data {
+    int id;
+    GtkTextMark *mark;
+};
+
+#endif /* PidginImage_h */
